Most stuff now done.

+++ source/BinaryAssetBuilder.Utility/BinaryAssetBuilder/Native/SizeT.cs
@@ -12,7 +12,7 @@ namespace BinaryAssetBuilder.Native
         private readonly ulong _value;
 #endif
 
-        public SizeT(ushort size)
+        public SizeT(uint size)
         {
             _value = size;
         }
